[refactor] CarStatus 합치기
