--- src/main/java/com/jaeyeonling/racingcar/domain/CarStatus.java
@@ -1,23 +1,0 @@
-package com.jaeyeonling.racingcar.domain;
-
-import com.jaeyeonling.racingcar.utils.StringUtils;
-
-public class CarStatus implements Visualizable {
-
-    static final String VISUAL_POSITION_STRING = "-";
-
-    private int position;
-
-    CarStatus(final Car car) {
-        position = car.getPosition();
-    }
-
-    @Override
-    public String visualize() {
-        return getVisualPosition();
-    }
-
-    private String getVisualPosition() {
-        return StringUtils.repeat(VISUAL_POSITION_STRING, position);
-    }
-}
